+++ module_ctrlproto/src/ctrlproto.h
@@ -24,20 +24,20 @@
  */
 typedef struct
 {
-	int operation_mode;			// 	Modes of Operation
-	int control_word;			// 	Control Word
+	int8_t  operation_mode;			// 	Modes of Operation
+	int16_t  control_word;			// 	Control Word
 
-	int target_torque;
-	int target_velocity;
-	int target_position;
+	int16_t  target_torque;
+	int32_t target_velocity;
+	int32_t target_position;
 
 
-	int operation_mode_display;	//	Modes of Operation Display
-	int status_word;			//  Status Word
+	int8_t operation_mode_display;	//	Modes of Operation Display
+	int16_t status_word;			//  Status Word
 
-	int torque_actual;
-	int velocity_actual;
-	int position_actual;
+	int16_t torque_actual;
+	int32_t velocity_actual;
+	int32_t position_actual;
 
 } ctrl_proto_values_t;
